requirements.txt
kucoin-python
PyYAML~=5.4.1
pandas~=1.5.3
ta-lib